requirements.txt
# Quant trading project dependencies
# Install with: pip install -r requirements.txt

# Core data processing
numpy==1.26.4
pandas==2.2.1
pandas-datareader==0.10.0
scipy==1.12.0
joblib==1.4.2

# Market data
yfinance==0.2.37
akshare==1.13.2
tushare==1.4.11
baostock==0.8.9

# Technical analysis
ta-lib==0.4.28
pandas-ta==0.3.14b0

# Visualization
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.20.0
pyecharts==2.0.5

# Backtesting and research
backtrader==1.9.78.123
zipline-reloaded==3.0.3
vectorbt==0.26.1
jupyter==1.0.0
ipython==8.22.2

# Machine learning
scikit-learn==1.4.1
xgboost==2.0.3
lightgbm==4.3.0

# Storage and services
sqlalchemy==2.0.28
pymongo==4.6.2
redis==5.0.3
fastapi==0.110.0
uvicorn==0.28.0
streamlit==1.32.2

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
schedule==1.2.1
loguru==0.7.2
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
pypdf==6.0.0

# Testing
pytest==8.1.1
pytest-cov==4.1.0
